# todo_swagger.yaml
openapi: 3.0.3
info:
  title: Swagger Todo
  version: "1.0.0"
servers:
  - url: http://localhost:8080/api
paths:
  /tasks/{id}:
    put:
      tags:
        - task
      summary: Update an existing task
      description: Update an existing task by Id
      operationId: updateTask
      parameters:
        - name: id
          in: path
          description: ID of task
          required: true
          schema:
            type: integer
            format: int
      requestBody:
        description: Update an existent task
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: number
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
  /tasks:
    post:
      tags:
        - task
      summary: Add a new task
      description: Add a new task
      operationId: CreateTask
      requestBody:
        description: Update an existent task
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: number
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
    get:
      tags:
        - task
      summary: Finds tasks
      operationId: findTasks
      parameters:
        - name: title
          in: query
          schema:
            type: string
        - name: description
          in: query
          schema:
            type: string
        - name: sort_by
          in: query
          schema:
            type: string
            enum:
              - title
              - status
              - created_at
              - updated_at
        - name: sort_order
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                  data:
                    type: array
                    items:
                      properties:
                        id:
                          type: number
                        title:
                          type: string
                        description:
                          type: string
                        created_at:
                          type: string
                          format: date-time
                        updated_at:
                          type: string
                          format: date-time
                        image:
                          type: string
                        status:
                          type: string
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
   
components:
  schemas:
    Task:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        image:
          type: string
          description: base64
        status:
          type: string
          enum:
            - IN_PROGRESS
            - COMPLETED